requirements.txt
coreapi==2.3.3
Django==3.1.3
django-environ==0.4.5
django-allauth==0.43.0
django-cors-headers==3.5.0
django-ckeditor==6.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.0.1
psycopg2-binary==2.8.6
